pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdn-scanner"
version = "0.1.0"
description = "CLI scanner for privacy-safe personal data discovery in mixed file stores"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "PDN Scanner Team" }
]
license = { text = "Proprietary" }
keywords = ["pii", "pdn", "scanner", "cli", "privacy"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Topic :: Security",
]
dependencies = [
  "typer>=0.12,<1.0",
  "pyyaml>=6.0,<7.0",
  "pydantic>=2.7,<3.0",
  "rich>=13.7,<14.0",
  "python-dotenv>=1.0,<2.0",
  "orjson>=3.10,<4.0",
  "beautifulsoup4>=4.12,<5.0",
  "lxml>=5.2,<6.0",
  "python-docx>=1.1,<2.0",
  "striprtf>=0.0.26,<1.0",
  "pypdf>=4.2,<5.0",
  "pdfplumber>=0.11,<1.0",
  "openpyxl>=3.1,<4.0",
  "xlrd>=2.0,<3.0",
  "pyarrow>=16.1,<18.0",
  "pandas>=2.2,<3.0",
  "pillow>=10.3,<12.0",
  "pytesseract>=0.3.10,<1.0",
  "charset-normalizer>=3.3,<4.0",
  "python-magic>=0.4.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
]

[project.scripts]
pdn-scan = "pdn_scanner.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
